Removing prod envs; confusing to maintain here
Since `alchemiscale-fah` is a component that *may* be deployed as part
of an `alchemiscale` deployment, it doesn't currently make much sense to
feature prod envs dedicated to deploying it. Instead, we should
encourage users of this package to deploy it as part of their own
`alchemiscale` environment configurations maintained elsewhere.

An example of this is
[alchemiscale.org-deployments](https://github.com/OpenFreeEnergy/alchemiscale.org-deployment).